moved movies into its own module
